fix: strengthen drop error handling
We previously just assumed any and all possible errors would be
EPERM, which, obviously, isn't the case. In setuid, we may get
EAGAIN, at which point we should spin on the syscall, and in both
setuid and setgid, we have to handle EINVAL.

Note that Unexpected_Errno is deliberatly uncaught in run - that
would be a very ugly failure indeed, and should look the part.